+++ pkg/interface/src/views/components/leap/Omnibox.js
