+++ pkg/interface/src/views/components/leap/OmniboxInput.js
